ci: success in forked repository (#670)
There're two workflows in forked repository that will fail: `docker-image.yml` and `release.yml`.

`docker-image.yml` fails because the secrets it needs does not exist in forked repository. Running it only in the `jeessy2/ddns-go` repository to fix this.

`release.yml` fails because the default `GITHUB_TOKEN` permission is now read-only. Add `contents: write` to fix this.

References:
- https://docs.github.com/actions/using-workflows/workflow-syntax-for-github-actions#example-only-run-job-for-specific-repository
- https://github.blog/changelog/2023-02-02-github-actions-updating-the-default-github_token-permissions-to-read-only/
- https://github.com/goreleaser/goreleaser-action#environment-variables