chore(dns): log only the root domain if not found in provider
Now, when we list a zone, the logs alone can't determine which parts of the domain are being used as a root domain.
Therefore, if `publicsuffix` provides a wrong root domain, the domain won't be found in the DNS provider even if it exists.

We can solve this by logging only the root domain, since only the root domain is used to list zones; users will also know what's causing it.